Add Notebooks Relevant to Lesson Plan
The contents of this commit are part of a Python course that can
be found [here](https://github.com/ehmatthes/intro_programming)

These are Juypter notebooks that we intend to modify and place
on our own personal Juypter server.  The notebooks included in
this commit cover the lesson plan we intend to teach for the
remote camp.