Remove breadcrumb links and Python 2 references. Set to 3.6.1
The breadcrumb links from the orignal versions of these notebooks
linked back to the orignal author's website.  Since we are using
our own host for these modified notebooks and the entire course
will not be used, we have decided to remove the links.

Addtionally, we have removed all Python 2 references from the
notebooks as our course will focus on Python 3.

The notebooks have also been set to target Python version 3.6.1
as that is the version our host is using.